+++ lib/elasticity/job_flow_status.rb
@@ -9,6 +9,7 @@ module Elasticity
     attr_accessor :created_at
     attr_accessor :started_at
     attr_accessor :ready_at
+    attr_accessor :ended_at
     attr_accessor :instance_count
     attr_accessor :master_instance_type
     attr_accessor :slave_instance_type
@@ -40,11 +41,14 @@ module Elasticity
 
       jobflow.created_at = Time.parse(xml_element.xpath('./ExecutionStatusDetail/CreationDateTime').text.strip)
 
+      ready_at = xml_element.xpath('./ExecutionStatusDetail/ReadyDateTime').text.strip
+      jobflow.ready_at = (ready_at == '') ? (nil) : (Time.parse(ready_at))
+
       started_at = xml_element.xpath('./ExecutionStatusDetail/StartDateTime').text.strip
       jobflow.started_at = (started_at == '') ? (nil) : (Time.parse(started_at))
 
-      ready_at = xml_element.xpath('./ExecutionStatusDetail/ReadyDateTime').text.strip
-      jobflow.ready_at = (ready_at == '') ? (nil) : (Time.parse(ready_at))
+      ended_at = xml_element.xpath('./ExecutionStatusDetail/EndDateTime').text.strip
+      jobflow.ended_at = (ended_at == '') ? (nil) : (Time.parse(ended_at))
 
       jobflow.instance_count = xml_element.xpath('./Instances/InstanceCount').text.strip
       jobflow.master_instance_type = xml_element.xpath('./Instances/MasterInstanceType').text.strip